# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestión de Inventario y Ventas
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ProfitFirst {
  operacion: number;
  dueno: number;
  ganancia: number;
}

export interface DashboardSummary {
  totalIncome: number;
  totalCost: number;
  totalShipping: number;
  fondoReposicion: number;
  netProfit: number;
  distributableProfit: number;
  profitFirst: ProfitFirst;
  totalSales: number;
  totalClients: number;
  lowStockCount: number;
}

export interface SalesChartPoint {
  month: string;
  income: number;
  profit: number;
}

export interface TopProduct {
  name: string;
  category: string;
  revenue: number;
  unitsSold: number;
}

export interface PerfumeryItem {
  id: number;
  name: string;
  brand: string;
  ml: number;
  stock: number;
  costPrice: number;
  salePrice: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreatePerfumeryItemBody {
  name: string;
  brand: string;
  ml: number;
  stock: number;
  costPrice: number;
  salePrice: number;
  /** @nullable */
  description?: string | null;
}

export interface UpdatePerfumeryItemBody {
  name?: string;
  brand?: string;
  ml?: number;
  stock?: number;
  costPrice?: number;
  salePrice?: number;
  /** @nullable */
  description?: string | null;
}

export type SublimationItemItemType =
  (typeof SublimationItemItemType)[keyof typeof SublimationItemItemType];

export const SublimationItemItemType = {
  maquinaria: "maquinaria",
  consumible: "consumible",
} as const;

export interface SublimationItem {
  id: number;
  name: string;
  category: string;
  itemType: SublimationItemItemType;
  /** @nullable */
  stock?: number | null;
  costPrice: number;
  salePrice: number;
  /** @nullable */
  description?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateSublimationItemBodyItemType =
  (typeof CreateSublimationItemBodyItemType)[keyof typeof CreateSublimationItemBodyItemType];

export const CreateSublimationItemBodyItemType = {
  maquinaria: "maquinaria",
  consumible: "consumible",
} as const;

export interface CreateSublimationItemBody {
  name: string;
  category: string;
  itemType: CreateSublimationItemBodyItemType;
  /** @nullable */
  stock?: number | null;
  costPrice: number;
  salePrice: number;
  /** @nullable */
  description?: string | null;
}

export type UpdateSublimationItemBodyItemType =
  (typeof UpdateSublimationItemBodyItemType)[keyof typeof UpdateSublimationItemBodyItemType];

export const UpdateSublimationItemBodyItemType = {
  maquinaria: "maquinaria",
  consumible: "consumible",
} as const;

export interface UpdateSublimationItemBody {
  name?: string;
  category?: string;
  itemType?: UpdateSublimationItemBodyItemType;
  /** @nullable */
  stock?: number | null;
  costPrice?: number;
  salePrice?: number;
  /** @nullable */
  description?: string | null;
}

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  city: string;
  department: string;
  /** @nullable */
  address?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClientBody {
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  city: string;
  department: string;
  /** @nullable */
  address?: string | null;
}

export interface UpdateClientBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  city?: string;
  department?: string;
  /** @nullable */
  address?: string | null;
}

export type SaleProductType =
  (typeof SaleProductType)[keyof typeof SaleProductType];

export const SaleProductType = {
  perfumeria: "perfumeria",
  sublimacion: "sublimacion",
} as const;

export interface Sale {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  productType: SaleProductType;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  costPrice: number;
  shippingCost: number;
  totalAmount: number;
  netProfit: number;
  /** @nullable */
  notes?: string | null;
  saleDate: string;
  createdAt: string;
}

export type CreateSaleBodyProductType =
  (typeof CreateSaleBodyProductType)[keyof typeof CreateSaleBodyProductType];

export const CreateSaleBodyProductType = {
  perfumeria: "perfumeria",
  sublimacion: "sublimacion",
} as const;

export interface CreateSaleBody {
  /** @nullable */
  clientId?: number | null;
  productType: CreateSaleBodyProductType;
  productId: number;
  quantity: number;
  unitPrice: number;
  /** @nullable */
  shippingCost?: number | null;
  /** @nullable */
  notes?: string | null;
  saleDate: string;
}

export interface UpdateSaleBody {
  /** @nullable */
  clientId?: number | null;
  quantity?: number;
  unitPrice?: number;
  costPrice?: number;
  shippingCost?: number;
  /** @nullable */
  notes?: string | null;
  saleDate?: string;
}

export interface TopClient {
  /** @nullable */
  clientId?: number | null;
  clientName: string;
  totalPurchases: number;
  salesCount: number;
}

export interface CategoryBreakdown {
  category: string;
  totalSales: number;
  totalRevenue: number;
}

export interface MonthlyReport {
  month: number;
  year: number;
  totalIncome: number;
  totalCost: number;
  netProfit: number;
  profitFirst: ProfitFirst;
  totalSales: number;
  sales: Sale[];
  topClients: TopClient[];
  categoryBreakdown: CategoryBreakdown[];
}

export interface Expense {
  id: number;
  amount: number;
  category: string;
  /** @nullable */
  description?: string | null;
  expenseDate: string;
  createdAt: string;
}

export interface CreateExpenseBody {
  amount: number;
  category: string;
  /** @nullable */
  description?: string | null;
  expenseDate: string;
}

export interface MonthlyGoal {
  id: number;
  month: number;
  year: number;
  targetAmount: number;
  createdAt: string;
}

export interface SetMonthlyGoalBody {
  month: number;
  year: number;
  targetAmount: number;
}

export type GetSalesChartParams = {
  year?: number;
};

export type ListSalesParams = {
  month?: number;
  year?: number;
};

export type ListExpensesParams = {
  month?: number;
  year?: number;
};

export type GetMonthlyGoalParams = {
  month: number;
  year: number;
};

export type GetMonthlyReportParams = {
  month: number;
  year: number;
};