# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Gestión de Inventario y Ventas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/sales-chart:
    get:
      operationId: getSalesChart
      tags: [dashboard]
      summary: Get monthly sales chart data
      parameters:
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Monthly sales data for chart
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesChartPoint"
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Bad request
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: integer
        - name: year
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    SalesChartPoint:
      type: object
      properties:
        month:
          type: string
        income:
          type: number
        profit:
          type: number
      required:
        - month
        - income
        - profit
...
    Client:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        city:
          type: string
        department:
          type: string
        address:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - city
        - department
        - createdAt
        - updatedAt
    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        city:
          type: string
        department:
          type: string
        address:
          type: ["string", "null"]
      required:
        - name
        - city
        - department
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        city:
          type: string
        department:
          type: string
        address:
          type: ["string", "null"]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        amount:
          type: number
        category:
          type: string
        description:
          type: ["string", "null"]
        expenseDate:
          type: string
        createdAt:
          type: string
      required:
        - id
        - amount
        - category
        - expenseDate
        - createdAt
    CreateExpenseBody:
      type: object
      properties:
        amount:
          type: number
        category:
          type: string
        description:
          type: ["string", "null"]
        expenseDate:
          type: string
      required:
        - amount
        - category
        - expenseDate
...