# lib/api-zod/src/generated/types/updateClientBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for Gestión de Inventario y Ventas
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateClientBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  city?: string;
  department?: string;
  /** @nullable */
  address?: string | null;
}